feat: seperate events to files, fix license, add getSelector function
