+++ packages/nexys/package.json
@@ -1,6 +1,6 @@
 {
   "name": "nexys",
-  "version": "1.1.4",
+  "version": "1.1.5",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "license": "Apache 2.0",
