+++ packages/nexys/src/utils/index.ts
@@ -27,6 +27,8 @@ export {
   collectVercelEnv,
   collectDOMData,
 } from "./collect.js";
+export { getSelector } from "./getSelector.js";
+export { getPagePath } from "./getPagePath.js";
 export const defaultOptions = {
   localStorage: {
     useLocalStorage: true,
@@ -36,6 +38,6 @@ export const defaultOptions = {
     testKey: "__nex-t__",
   },
   errors: {
-    allowAutomaticHandling: true, // Used for automatic exception handling.
+    allowAutomaticHandling: true,
   },
 };
